pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.15",
    "numpy>=2.2.2",
    "openai>=1.59.9",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "python-telegram-bot>=21.10",
    "scipy>=1.15.1",
    "tiktoken>=0.8.0",
]
